BackEnd/requirements.txt
fastapi[all]
sqlalchemy>=2.0.0
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pandas
scikit-learn
joblib
numpy